split code snippets into "for copy" and "context"
context (repeated code from previous snippets) is helpful, but it's most
efficient to use the "copy" button rendered into the corner of the snippet, and
if the context is included in the same code block as the snippet that's
intended to be copied, the context copied too and has to be deleted by the
reader/user. therefore breaking the context into it's own code block(s) makes
things more efficient for the user.