defer ITeleporterReceiver inheritance
including the inheritance at the beinning of the guide, but avoiding
implementing its methods until later in the guide, means that the user is going
through a good portion of the guide with an outstanding compiler error
("contract should be marked abstract [because it doesn't implement all the
methods of the inherited interfaces]"). it's better if the user never sees any
compiler errors.